% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{calc_astro}
\alias{calc_astro}
\title{Calculate astrophysical quantities}
\usage{
calc_astro(bestfit, highout, newDust = TRUE, specz = FALSE, cores = 1)
}
\arguments{
\item{bestfit}{ProSpect SED/bestfit object. Requires all the relevant
speclib, AGN, Dale templates}

\item{highout}{Highlander fit object with Posterior chains}

\item{newDust}{Whether to use metallicity dependent DTH/DTG and qPAH
from RR14 and Shivaei+24, boolean}

\item{specz}{Whether the fitted galaxy had a spectroscopy redshift
in which case only the specz is added to the table and not any uncertainty
ranges, boolean}

\item{cores}{Number of cores to use to create posterior samples, integer}
}
\value{
List of summary statistics and the full astrophysics posterior chain
}
\description{
Compute the astrophyical quantities from the Highlander posterior chains
}
