% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{fitSEDObs}
\alias{fitSEDObs}
\title{fitSEDObs}
\usage{
fitSEDObs(out, doAGN = TRUE, doDust = TRUE)
}
\arguments{
\item{out}{ProSpect SED/bestfit object}

\item{doAGN}{include the AGN emission, boolean}

\item{doDust}{include the Dust emission, boolean}
}
\value{
Data-frame of observed SEDs in Janskys and wavelengths in Ang
}
\description{
Compute the observer frame fited SEDS
}
