% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{fitSEDRest}
\alias{fitSEDRest}
\title{fitSEDRes}
\usage{
fitSEDRest(out, doAGN = TRUE, doDust = TRUE)
}
\arguments{
\item{out}{ProSpect SED/bestfit object}

\item{doAGN}{include the AGN emission, boolean}

\item{doDust}{include the Dust emission, boolean}
}
\value{
Data-frame of restframe SEDs in Lsol/Ang and wavelengths in Ang
}
\description{
Compute the rest frame fited SEDS
}
