% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{ionisingEfficiencyAGN}
\alias{ionisingEfficiencyAGN}
\title{Ionising efficiency for AGN only}
\usage{
ionisingEfficiencyAGN(out)
}
\arguments{
\item{out}{ProSpect SED/bestfit object}
}
\value{
Ionising efficiency in Hz/erg
}
\description{
Ionising efficiency for AGN only
}
