% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{New_DustMass_vDTH}
\alias{New_DustMass_vDTH}
\title{Dust mass stuff from D'Silva+26}
\usage{
New_DustMass_vDTH(out, qPAH_VSG = 0.035)
}
\arguments{
\item{out}{ProSpect bestfit object, class ProSpectSED}

\item{qPAH_VSG, }{Fraction of mass in very small grains and PAH molecules, numeric}
}
\value{
The updated dust mass, numeric
}
\description{
This function recomputes dust masses
using the wavelength dependent DTH from D'Silva+26
}
