% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{NionEmissivity}
\alias{NionEmissivity}
\title{Nion LyC emissivities}
\usage{
NionEmissivity(out)
}
\arguments{
\item{out}{ProSpect SED/bestfit object}
}
\value{
Nion LyC emissivities in photons/s
}
\description{
Calculates the ionizing photon production rate by directly integrating the
Lyman continuum (LyC) emission below 912 Angstrom from a ProSpect SED object.
This will be zero if the escape fraction is 0.
Better to remake the ProSpect SED with all the parameters the same but
manually make escape fraction = 0.
}
\details{
Nion LyC emissivities
}
