% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{NionEmissivityAGN}
\alias{NionEmissivityAGN}
\title{Nion LyC emissivities for AGN only}
\usage{
NionEmissivityAGN(out)
}
\arguments{
\item{out}{ProSpect SED/bestfit object}
}
\value{
Nion LyC emissivities in photons/s
}
\description{
Nion LyC emissivities for AGN only
}
