\name{ProSpectUtils-package}
\alias{ProSpectUtils-package}
\alias{ProSpectUtils}
\docType{package}
\title{
ProSpectUtils
}
\description{
Helpful functions to extract meaningful astrophysical quantities from ProSpect.
}
\details{
\tabular{ll}{
Package: \tab ProSpectUtils\cr
Type: \tab Package\cr
Version: \tab 0.0.3\cr
License: \tab LGPL-3\cr
Depends: \tab R (>= 4.0.0)\cr
Imports: \tab celestial, ProSpect, ProSpectData\cr
}
}
\author{
Maintainer: Jordan D'Silva
}