% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{RR14_BPL}
\alias{RR14_BPL}
\title{Remy-Ruyer+14 metallicity dependent DTH/DTG}
\usage{
RR14_BPL(Z, doDTG = FALSE)
}
\arguments{
\item{Z}{metallicity NOT in solar units, numeric}

\item{doDTG, }{Whether to compute DTH or DTG which are related by the, boolean
mean atomic weight}
}
\value{
DTH/DTG ratios which should be compared to
DTH=0.0073 the default in ProSpect, numeric
}
\description{
This function produces the linear relation
between metallicity and the DTH/DTG ratios
from Remy-Ruyer+14
}
