% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{SFR10}
\alias{SFR10}
\title{SFR10}
\usage{
SFR10(out)
}
\arguments{
\item{out}{ProSpect SED/bestfit object}
}
\value{
SFR averaged over 10Myr in solar masses per year
}
\description{
SFR averaged over 10Myr
}
