% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{shivaei24_qPAHZ}
\alias{shivaei24_qPAHZ}
\title{Shivaei+24 metallicity dependent qPAH}
\usage{
shivaei24_qPAHZ(Z)
}
\arguments{
\item{Z}{metallicity NOT in solar units, numeric}
}
\value{
qPAH, mass fraction in very small grains and PAH
}
\description{
This function produces the piecewise relation
between metallicity and the qPAH
from Shivaei+24
}
