% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{stellar_mass}
\alias{stellar_mass}
\title{Stellar masses}
\usage{
stellar_mass(out)
}
\arguments{
\item{out}{ProSpect SED/bestfit object}
}
\value{
Stellar mass in solar masses
}
\description{
Stellar masss
}
