% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{unlog_parm}
\alias{unlog_parm}
\title{unlog_parm}
\usage{
unlog_parm(out)
}
\arguments{
\item{out}{ProSpect SED/bestfit object}
}
\value{
SFR averaged over 10Myr in solar masses per year
}
\description{
Get the unlogged parameters
}
