% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{UV_tophat}
\alias{UV_tophat}
\alias{UVLum1500}
\title{Top-hat filter on 1500 Ang}
\usage{
UV_tophat(wave)
}
\arguments{
\item{wave}{wavelengths in angstrom, numeric}
}
\value{
Throughput
}
\description{
Filter for computing UV1500 luminosities
}
