% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{UVLum1500}
\alias{UVLum1500}
\title{UV1500 luminosities}
\usage{
UVLum1500(out)
}
\arguments{
\item{out}{ProSpect SED/bestfit object}
}
\value{
Rest-frame UV1500 luminosity in MAB (10pc)
}
\description{
UV1500 luminosities
}
