% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{UVLumAGN1500}
\alias{UVLumAGN1500}
\title{UV1500 luminosities for AGN only}
\usage{
UVLumAGN1500(out)
}
\arguments{
\item{out}{ProSpect SED/bestfit object}
}
\value{
Rest-frame UV1500 luminosity in MAB (10pc)
}
\description{
UV1500 luminosities for AGN only
}
