% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProSpectUtils.R
\name{calc_astro}
\alias{calc_astro}
\title{Calculate astrophysical quantities}
\usage{
calc_astro(
  bestfit,
  highout,
  newDust = TRUE,
  specz = FALSE,
  user_func = function(x) {
     NULL
 },
  Npost = NULL,
  cores = 1
)
}
\arguments{
\item{bestfit}{ProSpect SED/bestfit object. Requires all the relevant
speclib, AGN, Dale templates}

\item{highout}{Highlander fit object with Posterior chains}

\item{newDust}{Whether to use metallicity dependent DTH/DTG and qPAH
from RR14 and Shivaei+24, boolean}

\item{specz}{Whether the fitted galaxy had a spectroscopy redshift
in which case only the specz is added to the table and not any uncertainty
ranges, boolean}

\item{user_func}{Extra function that takes in ProSpect SED/bestfit object and
returns named vector of quantities, function}

\item{Npost}{Number posterior samples to use from the first one, numeric}

\item{cores}{Number of cores to use to create posterior samples, integer}
}
\value{
List of summary statistics and the full astrophysics posterior chain
}
\description{
Compute the astrophyical quantities from the Highlander posterior chains
}
